Move all files to new structure
