+++ src/main/java/com/emorn/bettercables/objects/api/forge/blocks/connector/BlockConnector.java
@@ -5,9 +5,9 @@ import com.emorn.bettercables.init.BlockInit;
 import com.emorn.bettercables.objects.api.forge.blocks.cable.BlockCable;
 import com.emorn.bettercables.objects.api.forge.common.AxisAlignedBoundingBoxConverter;
 import com.emorn.bettercables.objects.api.forge.common.BaseCable;
+import com.emorn.bettercables.objects.api.forge.common.Direction;
 import com.emorn.bettercables.objects.application.blocks.cable.CableAxisAlignedBoundingBox;
 import com.emorn.bettercables.objects.application.blocks.connector.ConnectorAxisAlignedBoundingBox;
-import com.emorn.bettercables.objects.blocks.connector.*;
 import com.emorn.bettercables.utils.IHasModel;
 import com.emorn.bettercables.utils.Reference;
 import mcp.MethodsReturnNonnullByDefault;
