+++ src/main/java/com/emorn/bettercables/objects/api/forge/blocks/connector/ConnectorNetwork.java
@@ -1,4 +1,4 @@
-package com.emorn.bettercables.objects.blocks.connector;
+package com.emorn.bettercables.objects.api.forge.blocks.connector;
 
 import com.emorn.bettercables.objects.application.blocks.connector.ConnectorSide;
 import com.emorn.bettercables.objects.application.blocks.connector.PossibleSlots;
