+++ src/main/java/com/emorn/bettercables/objects/api/forge/blocks/connector/ContainerConnector.java
@@ -1,6 +1,5 @@
-package com.emorn.bettercables.objects.blocks.connector;
+package com.emorn.bettercables.objects.api.forge.blocks.connector;
 
-import com.emorn.bettercables.objects.api.forge.blocks.connector.TileEntityConnector;
 import mcp.MethodsReturnNonnullByDefault;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.entity.player.InventoryPlayer;
