+++ src/main/java/com/emorn/bettercables/objects/api/forge/blocks/connector/NetworkManager.java
@@ -1,8 +1,6 @@
-package com.emorn.bettercables.objects.blocks.connector;
+package com.emorn.bettercables.objects.api.forge.blocks.connector;
 
-import com.emorn.bettercables.objects.api.forge.blocks.connector.BlockConnector;
 import com.emorn.bettercables.objects.api.forge.blocks.cable.BlockCable;
-import com.emorn.bettercables.objects.api.forge.blocks.connector.TileEntityConnector;
 import mcp.MethodsReturnNonnullByDefault;
 import net.minecraft.block.Block;
 import net.minecraft.util.math.BlockPos;
